changed Application entity id type to BigInteger

+++ src/main/java/com/skysoft/vaultlogic/common/domain/application/ApplicationRepository.java
@@ -7,16 +7,17 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
 import java.util.Optional;
 
 @Repository
-public interface ApplicationRepository extends JpaRepository<Application, Long> {
+public interface ApplicationRepository extends JpaRepository<Application, BigInteger> {
 
     Optional<Application> findByName(String name);
 
     SmartContractApplication findSmartContractApplicationByName(String name);
 
-    Optional<AppUriStatus> findAppUriStatusById(Long id);
+    Optional<AppUriStatus> findAppUriStatusById(BigInteger id);
 
     @Query("select a.uri from Application a where a.name = :name")
     Optional<String> selectApplicationUrlByName(@Param("name") String name);
